Extract CLI classes to separate files
